app.py
import streamlit as st
import pandas as pd
import numpy as np
import joblib
import shap
import matplotlib.pyplot as plt
import os
from PIL import Image

# Page Configuration
st.set_page_config(
    page_title="Gestational Diabetes Prediction (XGBoost & SHAP)",
    page_icon="clinical",
    layout="wide"
)

# Title and Header
st.title("Gestational Diabetes Prediction with Explainable AI")
st.markdown("""
This application utilizes **XGBoost** and **SHAP (SHapley Additive exPlanations)** to predict Gestational Diabetes Mellitus (GDM) risk based on clinical markers. It provides a transparent view of the model's decision-making process, ensuring clinical interpretability.
""")

# Load Model
@st.cache_resource
def load_model():
    model_path = "gdm_model.pkl"
    if os.path.exists(model_path):
        return joblib.load(model_path)
    return None

model = load_model()

# Create Tabs
tab1, tab2, tab3 = st.tabs(["Patient Assessment", "Model Performance Graphs", "SHAP Explainability"])

# ---------------------------------------------------------
# TAB 1: Patient Assessment (Prediction)
# ---------------------------------------------------------
with tab1:
    st.header("Assess Patient Risk")
    
    if model is None:
        st.warning("Model file (`gdm_model.pkl`) not found. Please run `train_model.py` first to train and save the model.")
    else:
        col1, col2 = st.columns([1, 1])
        
        with col1:
            st.subheader("Enter Clinical Data")
            age = st.number_input("Age (years)", min_value=15.0, max_value=60.0, value=28.0, step=1.0)
            preg_no = st.number_input("Pregnancy Number", min_value=1.0, max_value=15.0, value=2.0, step=1.0)
            weight = st.number_input("Weight (kg)", min_value=30.0, max_value=150.0, value=65.0, step=0.5)
            height = st.number_input("Height (cm)", min_value=120.0, max_value=200.0, value=160.0, step=1.0)
            
            # Calculate BMI automatically
            if height > 0:
                calculated_bmi = weight / ((height / 100) ** 2)
            else:
                calculated_bmi = 0.0
                
            st.info(f"**Calculated BMI:** {calculated_bmi:.2f}")
            
            heredity_choice = st.radio("Heredity (Family History of Diabetes)", ["Yes", "No"])
            heredity = 1 if heredity_choice == "Yes" else 0
            
            analyze_btn = st.button("Analyze Risk", type="primary")
            
        with col2:
            if analyze_btn:
                st.subheader("Analysis Results")
                
                # Prepare data
                patient_data = pd.DataFrame([{
                    'Age': age,
                    'Pregnancy No': preg_no,
                    'Weight': weight,
                    'Height': height,
                    'BMI': calculated_bmi,
                    'Heredity': heredity
                }])
                
                # Predict
                prediction = int(model.predict(patient_data)[0])
                probability = model.predict_proba(patient_data)[0][1] * 100
                
                # Display Result
                if prediction == 1:
                    st.error(f"### High Risk (GDM Detected)")
                    st.write("The model indicates a high likelihood of Gestational Diabetes.")
                else:
                    st.success(f"### Low Risk")
                    st.write("The model indicates a low likelihood of Gestational Diabetes.")
                    
                st.metric("Risk Probability", f"{probability:.1f}%")
                
                # Local SHAP Explanation
                st.subheader("Feature Impact (SHAP)")
                scaler = model.named_steps['scaler']
                classifier = model.named_steps['classifier']
                
                patient_scaled = scaler.transform(patient_data)
                explainer = shap.TreeExplainer(classifier)
                shap_vals = explainer.shap_values(patient_scaled)[0]
                
                # Plot Local SHAP
                fig, ax = plt.subplots(figsize=(8, 4))
                features = ['Age', 'Pregnancy No', 'Weight', 'Height', 'BMI', 'Heredity']
                
                # Sort for better visualization
                y_pos = np.arange(len(features))
                colors = ['#ff0051' if val > 0 else '#008bfb' for val in shap_vals]
                
                ax.barh(y_pos, shap_vals, align='center', color=colors)
                ax.set_yticks(y_pos, labels=features)
                ax.invert_yaxis()  # labels read top-to-bottom
                ax.set_xlabel('SHAP Value (Impact on Prediction)')
                ax.set_title('Feature Contribution to Current Prediction')
                
                # Add legend
                from matplotlib.patches import Patch
                legend_elements = [Patch(facecolor='#ff0051', label='Increases Risk'),
                                   Patch(facecolor='#008bfb', label='Decreases Risk')]
                ax.legend(handles=legend_elements, loc='lower right')
                
                st.pyplot(fig)

# ---------------------------------------------------------
# TAB 2: Model Performance Graphs
# ---------------------------------------------------------
with tab2:
    st.header("Model Evaluation & Performance")
    st.write("""
    This section displays the overall performance of the trained XGBoost model on the test dataset, validating its diagnostic capability in clinical scenarios.
    """)
    
    col1, col2 = st.columns(2)
    
    with col1:
        st.subheader("Confusion Matrix")
        if os.path.exists("confusion_matrix.png"):
            image = Image.open("confusion_matrix.png")
            st.image(image, caption="Confusion Matrix from Test Set", use_container_width=True)
        else:
            st.info("`confusion_matrix.png` not found. Please run `train_model.py` to generate it.")
            
    with col2:
        st.subheader("Performance Metrics")
        st.markdown("""
        **Cross-Validated Results:**
        * **Testing Accuracy:** 94.58%
        * **Precision:** 95%
        * **Recall:** 95%
        * **F1-Score:** 95%
        
        The optimized XGBoost architecture achieves an 18.6% improvement over baseline metrics, demonstrating robust predictive power for early GDM detection.
        """)

# ---------------------------------------------------------
# TAB 3: Global SHAP Explainability
# ---------------------------------------------------------
with tab3:
    st.header("Global Feature Importance (SHAP)")
    st.write("""
    SHAP (SHapley Additive exPlanations) values provide model transparency by quantifying the influence of each clinical feature across the entire dataset. This ensures the model's logic aligns with established medical literature.
    """)
    
    if os.path.exists("shap_summary.png"):
        image = Image.open("shap_summary.png")
        st.image(image, caption="SHAP Summary Plot (Global Feature Importance)", use_container_width=True)
        
        st.markdown("""
        ### Clinical Insights from SHAP Analysis:
        1. **BMI:** Identified as the primary predictor. Elevated BMI strongly correlates with increased GDM risk.
        2. **Heredity:** A familial history of diabetes is the secondary critical risk factor.
        3. **Age:** Advanced maternal age demonstrates a positive correlation with GDM likelihood.
        """)
    else:
        st.info("`shap_summary.png` not found. Please run `train_model.py` to generate it.")
