requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
xgboost>=2.0.0
shap>=0.42.0
joblib>=1.3.0
streamlit>=1.30.0
matplotlib>=3.7.0
Pillow>=10.0.0
